clean up psychophysics vs samples organization
